requirements.txt
fuzzywuzzy
lexical-diversity
python-Levenshtein
sacrebleu
sacremoses
setuptools
